more alpha fixes
Took 12 minutes
