+++ src/pginter/widgets/_frame.py
@@ -7,10 +7,10 @@ Frame - the base widget
 Author:
 Nilusink
 """
+from ..utils import add_corners, pil_image_to_surface
 from ._geo_manager import GeometryManager
 from concurrent.futures import Future
 from ..theme import ThemeManager
-from ..utils import merge_alphas
 from ..types import *
 import typing as tp
 import pygame as pg
@@ -34,61 +34,6 @@ if tp.TYPE_CHECKING:
 RES_T = tp.TypeVar("RES_T")
 
 
-# utility functions
-def add_corners(
-        im: Image,
-        ulr: int,
-        urr: int,
-        llr: int,
-        lrr: int
-) -> Image:
-    """
-    adds corners to an image
-
-    :param im: input image
-    :param ulr: upper left radius
-    :param urr: upper right radius
-    :param llr: lower left radius
-    :param lrr: lower right radius
-    :returns: the converted image
-    """
-    alpha = Image.new('L', im.size, 255)
-    w, h = im.size
-
-    # upper left corner
-    ulc = Image.new("L", (ulr * 2, ulr * 2), 0)
-    draw = ImageDraw.Draw(ulc)
-    draw.ellipse((0, 0, ulr * 2 - 1, ulr * 2 - 1), fill=255)
-    ulc = ulc.crop((0, 0, ulr, urr))
-    alpha.paste(ulc, (0, 0))
-
-    # upper right corner
-    urc = Image.new("L", (urr * 2, urr * 2), 0)
-    draw = ImageDraw.Draw(urc)
-    draw.ellipse((0, 0, urr * 2 - 1, urr * 2 - 1), fill=255)
-    urc = urc.crop((urr, 0, 2 * urr, urr))
-    alpha.paste(urc, (w - urr, 0))
-
-    # lower left radius
-    llc = Image.new("L", (llr * 2, llr * 2), 0)
-    draw = ImageDraw.Draw(llc)
-    draw.ellipse((0, 0, llr * 2 - 1, llr * 2 - 1), fill=255)
-    llc = llc.crop((0, llr, llr, llr * 2))
-    alpha.paste(llc, (0, h - llr))
-
-    # lower right radius
-    lrc = Image.new("L", (lrr * 2, lrr * 2), 0)
-    draw = ImageDraw.Draw(lrc)
-    draw.ellipse((0, 0, lrr * 2 - 1, lrr * 2 - 1), fill=255)
-    lrc = lrc.crop((lrr, lrr, lrr * 2, lrr * 2))
-    alpha.paste(lrc, (w - lrr, h - lrr))
-
-    # apply the corner radius, keeping the image's alpha (if already present)
-    *_, ia = im.split()  # extract alpha channel from image
-    im.putalpha(merge_alphas(alpha, ia, min))
-    return im
-
-
 def display_configurify(key: str) -> str:
     """
     convert a Frame init key to it's corresponding DisplayConfig key
@@ -106,17 +51,6 @@ def display_configurify(key: str) -> str:
             return config
 
     return key
-
-
-def pil_image_to_surface(pil_image):
-    """
-    helper function for converting pillow images to pygame images
-    """
-    return pg.image.fromstring(
-        pil_image.tobytes(),
-        pil_image.size,
-        pil_image.mode
-    ).convert_alpha()
 
 
 class _Bind(tp.TypedDict):
@@ -461,19 +395,6 @@ class Frame(GeometryManager):
         match event:
             case ThemeManager.NotifyEvent.theme_reload:
                 print("theme changed")
-                # the theme has been reloaded
-                # if not self._display_config_configured.bg:
-                #     if isinstance(self.__parent, Frame) and self.__parent._display_config["bg"] == self.theme.frame.bg1:
-                #         self._display_config.bg = self.theme.frame.bg2
-                #
-                #     else:
-                #         self._display_config.bg = self.theme.frame.bg1
-                #
-                # if not self._display_config_configured.border_color:
-                #     self._display_config.border_color = self.theme.frame.border
-                #
-                # if not self._display_config_configured.border_radius:
-                #     self._display_config.border_radius = self.theme.frame.border_radius
 
             case Style.NotifyEvent.property_change:
                 print(f"style changed: {info}, new value: {self.style[info]}")
